requirements.txt
python-decouple==3.0
requests==2.10.0
